Removes code generated with gowsdl
VMware vSphere WSDL is a gigantic service with
a flattened out list of methods. It is very
cumbersome for anybody to try to develop a decent
application using code generated with this WSDL.
Additionally it does not have documentation whatsoever.

A new approach has been taken in order to generate
a friendly and fully documented API.