Implements VSphereSession and converts returned dynamic property to exact type in function currentProperty()
For the dynamic property conversion I initially
did an implementation using the reflect package but ended up
settling with a closure in order to make it faster.

The registry map is necessary because Go is a compiled language
and its reflect packagemdoes not have any  means to get a Type,
in runtime, based on a string containing its name.

VSphereSession was created and for now it only allows
to create a session per application. Further ahead
we could explore some changes to allow creating multiple
vsphere sessions per app.