Converts returned dynamic property to exact type in function currentProperty()
I did a first implementation using the reflect package but ended up
settling with a closure in order to make it faster.

The registry map is necessary because Go is a compiled language
and its reflect packagemdoes not have any  means to get a Type,
in runtime, based on a string containing its name.